build works.

+++ KFC_Cpp/src/Board.cpp
@@ -10,8 +10,8 @@ Board::Board(int cell_H_pix_,
              int W_cells_,
              int H_cells_,
              const ImgPtr& image,
-             float cell_H_m_,
-             float cell_W_m_)
+             double cell_H_m_,
+             double cell_W_m_)
     : cell_H_pix(cell_H_pix_),
       cell_W_pix(cell_W_pix_),
       W_cells(W_cells_),
@@ -32,23 +32,23 @@ void Board::show() const {
 }
 
 // ---------------------------------------------------------------------------
-std::pair<int, int> Board::m_to_cell(const std::pair<float, float>& pos_m) const {
-    float x_m = pos_m.first;
-    float y_m = pos_m.second;
+std::pair<int, int> Board::m_to_cell(const std::pair<double, double>& pos_m) const {
+    double x_m = pos_m.first;
+    double y_m = pos_m.second;
     int col = static_cast<int>(std::round(x_m / cell_W_m));
     int row = static_cast<int>(std::round(y_m / cell_H_m));
     return {row, col};
 }
 
-std::pair<float, float> Board::cell_to_m(const std::pair<int, int>& cell) const {
+std::pair<double, double> Board::cell_to_m(const std::pair<int, int>& cell) const {
     int r = cell.first;
     int c = cell.second;
-    return {static_cast<float>(c) * cell_W_m, static_cast<float>(r) * cell_H_m};
+    return {static_cast<double>(c) * cell_W_m, static_cast<double>(r) * cell_H_m};
 }
 
-std::pair<int, int> Board::m_to_pix(const std::pair<float, float>& pos_m) const {
-    float x_m = pos_m.first;
-    float y_m = pos_m.second;
+std::pair<int, int> Board::m_to_pix(const std::pair<double, double>& pos_m) const {
+    double x_m = pos_m.first;
+    double y_m = pos_m.second;
     int x_px = static_cast<int>(std::round(x_m / cell_W_m * cell_W_pix));
     int y_px = static_cast<int>(std::round(y_m / cell_H_m * cell_H_pix));
     return {x_px, y_px};
