+++ KFC_Cpp/src/Board.hpp
@@ -11,8 +11,8 @@ public:
     int H_cells;
 
     ImgPtr img;              // board image
-    float cell_H_m;
-    float cell_W_m;
+    double cell_H_m;
+    double cell_W_m;
 
 public:
     Board(int cell_H_pix,
@@ -20,8 +20,8 @@ public:
           int W_cells,
           int H_cells,
           const ImgPtr& image,
-          float cell_H_m = 1.0f,
-          float cell_W_m = 1.0f);
+          double cell_H_m = 1.0f,
+          double cell_W_m = 1.0f);
 
     Board(const Board&) = default;
     Board(Board&&) noexcept = default;
@@ -33,8 +33,8 @@ public:
     void show() const;                   // Show only if an image is loaded
 
     // Coordinate conversions -------------------------------------------------
-    std::pair<int, int> m_to_cell(const std::pair<float, float>& pos_m) const;
-    std::pair<float, float> cell_to_m(const std::pair<int, int>& cell) const;
-    std::pair<int, int> m_to_pix(const std::pair<float, float>& pos_m) const;
+    std::pair<int, int> m_to_cell(const std::pair<double, double>& pos_m) const;
+    std::pair<double, double> cell_to_m(const std::pair<int, int>& cell) const;
+    std::pair<int, int> m_to_pix(const std::pair<double, double>& pos_m) const;
 
 };
